Separate renderer entry points from app.
Renderer entry points are specific to the desktop build of studio.
Move the electron specific entry elements to desktop/renderer.

To avoid top-level directory confusion, desktop now contains: main,
preload, and renderer. Each has the code for their equivalent electron
space. The webpack configurations are moved from the top level into desktop
since they are for the desktop build.

App exposes a "webpack" file with a function to build a partial webpack
configuration required to integrate app into renderer.

App exposes a "typings" file with global type information assumed by app
code and required to integrate app into renderer.
