+++ app/typings/index.d.ts
@@ -1,0 +1,11 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+// Custom types available globally and required when using app components as dependencies
+
+import "./extensions";
+import "./react";
+import "./fluentui";
+import "./styled-components";
+import "./webpack-defines";
