+++ desktop/main/.eslintrc.yaml
@@ -1,8 +1,3 @@
-settings:
-  import/resolver:
-    webpack:
-      config: ./webpack.main.config.ts
-
 rules:
   # Re-enable these when https://github.com/benmosher/eslint-plugin-import/issues/1996 is fixed
   import/no-unresolved: off
