+++ desktop/main/StudioWindow.ts
