+++ desktop/main/rosPackageResources.ts
