+++ desktop/main/simulateUserClick.ts
