+++ desktop/main/telemetry.ts
