+++ desktop/preload/LocalFileStorage.ts
