+++ jest.config.json
@@ -1,8 +1,14 @@
 {
-  "//": "Our top level jest config to reference all the projects under our monorepo. We exclude the integration test since it involves performing a webpack build and doesn't support 'watch' or 'debug' in the same way",
+  "//": "Our top level jest config to reference all the projects under our monorepo. We exclude the desktop test since it involves performing a webpack build and doesn't support 'watch' or 'debug' in the same way",
   "projects": [
+    "<rootDir>/jest.config.json",
     "<rootDir>/app/jest.config.json",
     "<rootDir>/ci/jest.config.json",
     "<rootDir>/packages/*/jest.config.json"
-  ]
+  ],
+  "testRunner": "jest-circus/runner",
+  "transform": {
+    "\\.[jt]sx?$": ["babel-jest", { "rootMode": "upward" }]
+  },
+  "testMatch": ["**/licenses.test.ts"]
 }
