+++ package.json
@@ -18,17 +18,17 @@
   "scripts": {
     "clean": "tsc --build --clean packages/*/tsconfig.json && rimraf .webpack dist app/storybook-screenshots app/storybook-static",
     "clean:package": "yarn && yarn clean && yarn build:prod && yarn package",
-    "start": "electron .webpack",
+    "start": "electron desktop/.webpack",
     "build:packages": "tsc --build --verbose packages/*/tsconfig.json",
-    "build:prod": "webpack --mode production --progress",
-    "build:dev": "webpack --mode development --progress",
-    "serve": "webpack serve --mode development --progress",
+    "build:prod": "webpack --mode production --progress --config desktop/webpack.config.ts",
+    "build:dev": "webpack --mode development --progress --config desktop/webpack.config.ts",
+    "serve": "webpack serve --mode development --progress --config desktop/webpack.config.ts",
     "lint": "TIMING=1 eslint --rulesdir eslint-rules --report-unused-disable-directives --fix .",
     "lint:ci": "TIMING=1 NODE_OPTIONS='--max-old-space-size=4096' eslint --rulesdir eslint-rules --report-unused-disable-directives --config .eslintrc.ci.yaml .",
     "test": "jest",
     "test:debug": "NODE_OPTIONS='--inspect-brk' jest",
     "test:watch": "jest --watch",
-    "test:integration": "jest --forceExit --projects test",
+    "test:integration": "jest --forceExit --projects desktop",
     "package:win": "yarn package --win",
     "package:darwin": "yarn package --mac",
     "package:linux": "yarn package --linux",
