+++ tsconfig.json
@@ -9,7 +9,6 @@
     "./**/jest.config.ts",
     "./ci/**/*.ts",
     "./eslint-rules/*.js",
-    "./test/**/*.ts",
     "./resources/notarize.ts",
     "./package.json"
   ],
